grouped flag and arg functions into different file
